+++ src/main/java/dev/piste/vayna/apis/officer/gson/Queue.java
@@ -1,4 +1,4 @@
-package dev.piste.vayna.apis.valorantapi.gson;
+package dev.piste.vayna.apis.officer.gson;
 
 /**
  * @author Piste | https://github.com/zPiste
