+++ src/main/java/dev/piste/vayna/apis/officer/gson/agent/Ability.java
@@ -1,0 +1,28 @@
+package dev.piste.vayna.apis.officer.gson.agent;
+
+// GSON CLASS
+@SuppressWarnings("ALL")
+public class Ability {
+
+    private String slot;
+    private String displayName;
+    private String description;
+    private String displayIcon;
+
+    public String getSlot() {
+        return slot;
+    }
+
+    public String getDisplayName() {
+        return displayName;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public String getDisplayIcon() {
+        return displayIcon;
+    }
+
+}
