--- src/main/java/dev/piste/vayna/apis/valorantapi/gson/Buddy.java
@@ -1,18 +1,0 @@
-package dev.piste.vayna.apis.valorantapi.gson;
-
-// GSON CLASS
-@SuppressWarnings("ALL")
-public class Buddy {
-
-    private String displayName;
-    private String displayIcon;
-
-    public String getDisplayName() {
-        return displayName;
-    }
-
-    public String getDisplayIcon() {
-        return displayIcon;
-    }
-
-}
